+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## [Unreleased]
 
+- Support `backgroundColor` and `color` local directives ([#32](https://github.com/marp-team/marpit/pull/32))
 - Suppress confusable theme import when tweaking ([#33](https://github.com/marp-team/marpit/pull/33))
 
 ## v0.0.6 - 2018-05-29